\name{fpropcov}
\alias{fpropcov}
\title{fpropcov}
\description{Calculeaza Cov(aX+bY,cX+dY).}
\usage{fpropcov(X,Y,repCom,a,b,c,d)}

\arguments{
  \item{X}{Functia de masa a v.a. X.}
  \item{Y}{Functia de masa a v.a. Y.}
  \item{repCom}{Repartitia comuna a v.a X si Y.}
  \item{a,b,c,d}{Constante.}
}

\details{}
\value{Cov(aX+bY,cX+dY).}
