\name{frepcomgen}
\alias{frepcomgen}
\title{frepcomgen}
\description{
Generează un tabel cu repartiţia comună a v.a. X şi Y incompletă, dar ȋntr-o formă ȋn care poate fi completată ulterior.}
\usage{frepcomgen(n,m)}

\arguments{
  \item{n}{Numar de linii. Numar pozitiv.}
  \item{m}{Numar de coloane. Numar pozitiv.}
}
\details{}
\value{Tabel cu repartiţia comună a v.a. X şi Y incompletă.}
}
