\name{frepmarginal}
\alias{frepmarginal}
\title{frepmarginal}
\description{
 Construiește repartițiile marginale pentru X și Y pornind de la repartiția lor comună.}
\usage{frepmarginal(t)}

\arguments{
  \item{t}{Tabel repartitia comuna completa.}
}
\details{}
\value{Repartițiile marginale pentru X și Y.}
}
