\name{fvaop}
\alias{fvaop}
\title{fvaop}
\description{
 Calculeaza repartitia v.a. X +-* Y.}
\usage{fvaop <- function(X,Y,repCom,op)}

\arguments{
  \item{X}{Functia de repartitie a v.a X.}
  \item{Y}{Functia de repartitie a v.a X.}
  \item{repCom}{Repartitia comuna a v.a X si Y.}
  \item{op}{'+' , '-' , '*'}
}
\details{}
\value{Varianta v.a. X.}
