\name{fverind}
\alias{fverind}
\title{fverind}
\description{
 Verifica daca cele 2 variabile sunt independente}
\usage{fverind(X,Y,repCom)}

\arguments{
  \item{X}{Functia de repartitie a v.a X.}
  \item{Y}{Functia de repartitie a v.a Y.}
  \item{repCom}{Repartitia comuna a variabilelor X si Y.}
}
\details{}
\value{True/False daca varibilele X si Y sunt independente}
